+++ products/jbrowse-desktop/src/sessionModel/TrackMenu.ts
@@ -16,6 +16,7 @@ import SettingsIcon from '@mui/icons-material/Settings'
 import type { DesktopRootModel } from '../rootModel/rootModel'
 import type PluginManager from '@jbrowse/core/PluginManager'
 import type { BaseTrackConfig } from '@jbrowse/core/pluggableElementTypes'
+import type { MenuItem } from '@jbrowse/core/ui'
 import type {
   SessionWithDialogs,
   SessionWithDrawerWidgets,
@@ -31,15 +32,122 @@ export function DesktopSessionTrackMenuMixin(_pluginManager: PluginManager) {
   return types.model({}).views(self => ({
     /**
      * #method
+     * raw track actions (Settings, Copy, Delete, Index) without submenu wrapper
      */
-    getTrackActionMenuItems(trackConfig: BaseTrackConfig) {
+    getTrackActions(trackConfig: BaseTrackConfig): MenuItem[] {
       const session = self as SessionWithDialogs &
         SessionWithTracks &
         SessionWithDrawerWidgets
-      // Handle both MST models and frozen/plain objects
       const trackSnapshot = structuredClone(
         isStateTreeNode(trackConfig) ? getSnapshot(trackConfig) : trackConfig,
       )
+      return [
+        {
+          label: 'Settings',
+          onClick: () => {
+            session.editConfiguration(trackConfig)
+          },
+          icon: SettingsIcon,
+        },
+        {
+          label: 'Copy track',
+          onClick: () => {
+            const now = Date.now()
+            trackSnapshot.trackId += `-${now}`
+            if (trackSnapshot.displays) {
+              for (const d of trackSnapshot.displays) {
+                d.displayId += `-${now}`
+              }
+            }
+            trackSnapshot.name += ' (copy)'
+            trackSnapshot.category = undefined
+            session.addTrackConf(trackSnapshot)
+          },
+          icon: CopyIcon,
+        },
+        {
+          label: 'Delete track',
+          onClick: () => {
+            session.deleteTrackConf(trackConfig)
+          },
+          icon: DeleteIcon,
+        },
+        ...(isSupportedIndexingAdapter(trackSnapshot.adapter?.type)
+          ? [
+              {
+                label: trackSnapshot.textSearching
+                  ? 'Re-index track'
+                  : 'Index track',
+                onClick: () => {
+                  const rootModel = getParent<DesktopRootModel>(self)
+                  const { jobsManager } = rootModel
+                  const { trackId, assemblyNames, textSearching, name } =
+                    trackSnapshot
+                  const indexName = `${name}-index`
+                  // TODO: open jobs list widget
+                  jobsManager.queueJob({
+                    indexingParams: {
+                      attributes: textSearching?.indexingAttributes || [
+                        'Name',
+                        'ID',
+                      ],
+                      exclude: textSearching?.indexingFeatureTypesToExclude || [
+                        'CDS',
+                        'exon',
+                      ],
+                      assemblies: assemblyNames,
+                      tracks: [trackId],
+                      indexType: 'perTrack',
+                      timestamp: new Date().toISOString(),
+                      name: indexName,
+                    },
+                    name: indexName,
+                  })
+                },
+                icon: Indexing,
+              },
+            ]
+          : []),
+      ]
+    },
+
+    /**
+     * #method
+     * flattened menu items for use in hierarchical track selector
+     */
+    getTrackListMenuItems(trackConfig: BaseTrackConfig): MenuItem[] {
+      const session = self as SessionWithDialogs &
+        SessionWithTracks &
+        SessionWithDrawerWidgets & {
+          getTrackActions: (c: BaseTrackConfig) => MenuItem[]
+        }
+      return [
+        {
+          label: 'About track',
+          onClick: () => {
+            session.queueDialog(doneCallback => [
+              AboutDialog,
+              { config: trackConfig, session, handleClose: doneCallback },
+            ])
+          },
+          icon: InfoIcon,
+        },
+        ...session.getTrackActions(trackConfig),
+      ]
+    },
+
+    /**
+     * #method
+     */
+    getTrackActionMenuItems(
+      trackConfig: BaseTrackConfig,
+      extraTrackActions?: MenuItem[],
+    ): MenuItem[] {
+      const session = self as SessionWithDialogs &
+        SessionWithTracks &
+        SessionWithDrawerWidgets & {
+          getTrackActions: (c: BaseTrackConfig) => MenuItem[]
+        }
       return [
         {
           label: 'About track',
@@ -55,76 +163,11 @@ export function DesktopSessionTrackMenuMixin(_pluginManager: PluginManager) {
           type: 'subMenu' as const,
           label: 'Track actions',
           subMenu: [
-            {
-              label: 'Settings',
-              onClick: () => {
-                session.editConfiguration(trackConfig)
-              },
-              icon: SettingsIcon,
-            },
-            {
-              label: 'Copy track',
-              onClick: () => {
-                const now = Date.now()
-                trackSnapshot.trackId += `-${now}`
-                if (trackSnapshot.displays) {
-                  for (const d of trackSnapshot.displays) {
-                    d.displayId += `-${now}`
-                  }
-                }
-                trackSnapshot.name += ' (copy)'
-                trackSnapshot.category = undefined
-                session.addTrackConf(trackSnapshot)
-              },
-              icon: CopyIcon,
-            },
-            {
-              label: 'Delete track',
-              onClick: () => {
-                session.deleteTrackConf(trackConfig)
-              },
-              icon: DeleteIcon,
-            },
-            ...(isSupportedIndexingAdapter(trackSnapshot.adapter?.type)
-              ? [
-                  {
-                    label: trackSnapshot.textSearching
-                      ? 'Re-index track'
-                      : 'Index track',
-                    onClick: () => {
-                      const rootModel = getParent<DesktopRootModel>(self)
-                      const { jobsManager } = rootModel
-                      const { trackId, assemblyNames, textSearching, name } =
-                        trackSnapshot
-                      const indexName = `${name}-index`
-                      // TODO: open jobs list widget
-                      jobsManager.queueJob({
-                        indexingParams: {
-                          attributes: textSearching?.indexingAttributes || [
-                            'Name',
-                            'ID',
-                          ],
-                          exclude:
-                            textSearching?.indexingFeatureTypesToExclude || [
-                              'CDS',
-                              'exon',
-                            ],
-                          assemblies: assemblyNames,
-                          tracks: [trackId],
-                          indexType: 'perTrack',
-                          timestamp: new Date().toISOString(),
-                          name: indexName,
-                        },
-                        name: indexName,
-                      })
-                    },
-                    icon: Indexing,
-                  },
-                ]
-              : []),
+            ...session.getTrackActions(trackConfig),
+            ...(extraTrackActions || []),
           ],
         },
-        { type: 'divider' },
+        { type: 'divider' as const },
       ]
     },
   }))
