+++ products/jbrowse-react-linear-genome-view/src/createModel/createSessionModel.ts
@@ -15,6 +15,7 @@ import {
 import InfoIcon from '@mui/icons-material/Info'
 
 import type PluginManager from '@jbrowse/core/PluginManager'
+import type { MenuItem } from '@jbrowse/core/ui'
 import type { AbstractSessionModel } from '@jbrowse/core/util/types'
 import type { Instance } from '@jbrowse/mobx-state-tree'
 import type { LinearGenomeViewStateModel } from '@jbrowse/plugin-linear-genome-view'
@@ -137,7 +138,10 @@ export default function sessionModelFactory(pluginManager: PluginManager) {
       /**
        * #method
        */
-      getTrackActionMenuItems(config: any) {
+      getTrackActionMenuItems(
+        config: any,
+        extraTrackActions?: MenuItem[],
+      ): MenuItem[] {
         return [
           {
             label: 'About track',
@@ -149,6 +153,7 @@ export default function sessionModelFactory(pluginManager: PluginManager) {
             },
             icon: InfoIcon,
           },
+          ...(extraTrackActions || []),
         ]
       },
     }))
