+++ products/jbrowse-web/src/tests/CopyAndDelete.test.tsx
@@ -56,6 +56,7 @@ test('copy and delete track in admin mode', () => {
       (await findAllByTestId(/prerendered_canvas/, {}, delay))[0]!,
     )
     fireEvent.click(await findByTestId('track_menu_icon'))
+    fireEvent.click(await findByText('Track actions'))
     fireEvent.click(await findByText('Delete track'))
     await waitFor(() => {
       expect(view.tracks.length).toBe(0)
@@ -73,10 +74,17 @@ test('copy and delete reference sequence track disabled', () => {
     view.setNewView(0.05, 5000)
     const trackConf = getConf(assemblyManager.get('volvox')!, 'sequence')
 
-    // @ts-expect-error
-    const trackMenuItems = session.getTrackActionMenuItems(trackConf)
+    const trackMenuItems = session.getTrackActionMenuItems!(trackConf)
+    const trackActionsSubMenu = trackMenuItems.find(
+      item => 'label' in item && item.label === 'Track actions',
+    )
+    const trackActions =
+      trackActionsSubMenu && 'subMenu' in trackActionsSubMenu
+        ? trackActionsSubMenu.subMenu
+        : []
 
     // copy ref seq track disabled
+    // Note: htsTrackEntryMenu uses getTrackListMenuItems which is a flat list
     fireEvent.click(
       await findByTestId('htsTrackEntryMenu-Tracks,volvox_refseq', {}, delay),
     )
@@ -86,8 +94,20 @@ test('copy and delete reference sequence track disabled', () => {
     await waitFor(() => {
       expect(view.tracks.length).toBe(0)
     })
-    expect(trackMenuItems[2].disabled).toBe(true)
-    expect(trackMenuItems[3].disabled).toBe(true)
+    const copyTrackItem = trackActions.find(
+      item => 'label' in item && item.label === 'Copy track',
+    )
+    const deleteTrackItem = trackActions.find(
+      item => 'label' in item && item.label === 'Delete track',
+    )
+    expect(
+      copyTrackItem && 'disabled' in copyTrackItem && copyTrackItem.disabled,
+    ).toBe(true)
+    expect(
+      deleteTrackItem &&
+        'disabled' in deleteTrackItem &&
+        deleteTrackItem.disabled,
+    ).toBe(true)
   })
 }, 40000)
 
@@ -114,6 +134,7 @@ test('copy and delete track to session tracks', () => {
       (await findAllByTestId(/prerendered_canvas/, {}, delay))[0]!,
     )
     fireEvent.click(await findByTestId('track_menu_icon'))
+    fireEvent.click(await findByText('Track actions'))
     fireEvent.click(await findByText('Delete track'))
     await waitFor(() => {
       expect(view.tracks.length).toBe(0)
